Working on "Add to cart" use case
I am currently able to add items to the DB under a user record.
However, I still need to implement logic that determines whether or not
the product type is already in the cart. In this scenario, the quantity
must increase.

I also deleted the Cart pojo because, as of now, I don’t see it as
being necessary.

Note that the cart JSP does not exist yet, so adding an item to the
cart will result in an error message.